facade test and service tests

+++ service/src/main/java/cz/muni/fi/service/DriverServiceImpl.java
@@ -28,6 +28,9 @@ public class DriverServiceImpl implements DriverService {
 
     @Override
     public Driver findDriverById(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Id argument cannot be null!");
+        }
         try{
             return driverDao.findDriverById(id);
         } catch(Exception ex){
@@ -37,6 +40,9 @@ public class DriverServiceImpl implements DriverService {
 
     @Override
     public Driver findDriverByName(String name, String surname) {
+        if (name == null || surname == null) {
+            throw new IllegalArgumentException("name or surname cannot be null!");
+        }
         try{
             return driverDao.findDriverByName(name, surname);
         } catch(Exception ex){
@@ -46,6 +52,9 @@ public class DriverServiceImpl implements DriverService {
 
     @Override
     public Driver findTestDriver(String name, String surname) {
+        if (name == null || surname == null) {
+            throw new IllegalArgumentException("name or surname cannot be null!");
+        }
         try{
             return driverDao.findTestDriver(name, surname);
         } catch(Exception ex){
@@ -55,6 +64,9 @@ public class DriverServiceImpl implements DriverService {
 
     @Override
     public void addDriver(Driver driver) {
+        if (driver == null) {
+            throw new IllegalArgumentException("driver argument cannot be null!");
+        }
         try{
             driverDao.addDriver(driver);
         } catch(Exception ex){
@@ -64,6 +76,9 @@ public class DriverServiceImpl implements DriverService {
 
     @Override
     public void updateDriver(Driver driver) {
+        if (driver == null) {
+            throw new IllegalArgumentException("driver argument cannot be null!");
+        }
         try{
             driverDao.updateDriver(driver);
         } catch(Exception ex){
@@ -73,6 +88,9 @@ public class DriverServiceImpl implements DriverService {
 
     @Override
     public void deleteDriver(Driver driver) {
+        if (driver == null) {
+            throw new IllegalArgumentException("driver argument cannot be null!");
+        }
         try{
             driverDao.deleteDriver(driver);
         } catch(Exception ex){
